+++ backend/package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.772.0",
     "@aws-sdk/s3-request-presigner": "^3.772.0",
-    "@types/stopword": "^2.0.3",
     "axios": "^1.8.2",
     "bcryptjs": "^3.0.2",
     "cors": "^2.8.5",
@@ -23,10 +22,8 @@
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.12.1",
     "multer": "^1.4.5-lts.1",
-    "natural": "^8.0.1",
     "nodemailer": "^6.10.0",
     "resend": "^4.1.2",
-    "stopword": "^3.1.4",
     "ts-node": "^10.9.2",
     "tsx": "^4.19.3",
     "uuid": "^11.1.0",
