+++ backend/src/middleware/auth.ts
@@ -13,13 +13,19 @@ declare global {
 }
 
 export const authenticateToken = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    // //TODO: Hardcoded user id for now
-
-    const authHeader = req.headers['authorization'];
-    const token = authHeader && authHeader.split(' ')[1];
-
-    req.user = { _id: token ?? "66e66e6e6e6e6e6e6e6e6e6e" };
+    //TODO: Hardcoded user id for now
+    req.user = { _id: "66e66e6e6e6e6e6e6e6e6e6e" };
     next();
+    // const authHeader = req.headers['authorization'];
+    // const token = authHeader && authHeader.split(' ')[1];
+
+    // if (!token) {
+    //     res.status(401).json({
+    //         success: false,
+    //         error: 'Access token required',
+    //     });
+    //     return;
+    // }
 
     // try {
     //     const decoded = jwt.verify(token, process.env.JWT_SECRET || 'your-secret-key') as { _id: string };
