+++ backend/src/services/blogService.ts
@@ -2,8 +2,6 @@ import mongoose, { Types } from 'mongoose';
 import { z } from 'zod';
 import Media from '../models/mediaModel.js';
 import Post from '../models/postModel.js';
-import contentSimilarityService from './contentSimilarityService.ts';
-import Product from '../models/productModel.ts';
 
 // Validation schemas
 const createPostSchema = z.object({
@@ -17,7 +15,7 @@ const createPostSchema = z.object({
 const updatePostSchema = z.object({
   title: z.string().min(1, 'Title is required').max(100, 'Title is too long').optional(),
   content: z.string().min(1, 'Content is required').optional(),
-  thumbnail: z.string().optional().nullable(),
+  thumbnail: z.string().optional(),
   category: z.string().min(1, 'Category is required').optional(),
   isPublished: z.boolean().optional(),
 });
@@ -28,11 +26,6 @@ class BlogService {
     try {
       // Validate input data
       const validatedData = createPostSchema.parse(postData);
-
-      //remove thumbnail if it is null
-      if (validatedData.thumbnail === null) {
-        delete validatedData.thumbnail;
-      }
 
       // Create new post
       const post = new Post({
@@ -54,13 +47,16 @@ class BlogService {
   async getAllPosts(page: number = 1, limit: number = 10, filters: any = {}) {
     try {
       const query: any = {};
-
+      
       // Apply filters
       if (filters.category) {
         query.category = filters.category;
       }
-      if(filters.search) {
-        query.title = filters.search;
+      if (filters.isPublished !== undefined) {
+        query.isPublished = filters.isPublished;
+      }
+      if (filters.author) {
+        query.author = new Types.ObjectId(filters.author);
       }
 
       const posts = await Post.find(query)
@@ -117,7 +113,6 @@ class BlogService {
 
   // Update a post
   async updatePost(postId: string, updateData: any, userId: string) {
-    console.log(updateData, postId, userId)
     try {
       if (!mongoose.Types.ObjectId.isValid(postId)) {
         throw new Error('Invalid post ID');
@@ -138,21 +133,7 @@ class BlogService {
       }
 
       // Update post
-      if (validatedData.title) {
-        post.title = validatedData.title;
-      }
-      if (validatedData.content) {
-        post.content = validatedData.content;
-      }
-      if (validatedData.category) {
-        post.category = validatedData.category;
-      }
-      if (validatedData.thumbnail) {
-        post.thumbnail = validatedData.thumbnail;
-      }
-      if (validatedData.isPublished) {
-        post.isPublished = validatedData.isPublished;
-      }
+      Object.assign(post, validatedData);
       await post.save();
 
       return post;
@@ -160,7 +141,6 @@ class BlogService {
       if (error instanceof z.ZodError) {
         throw new Error(`Validation error: ${error.errors[0].message}`);
       }
-      console.log(error)
       throw new Error('Failed to update post');
     }
   }
@@ -210,7 +190,7 @@ class BlogService {
 
       const userObjectId = userId;
       const likeIndex = post.likes.findIndex(like => like === userObjectId);
-
+      
       if (likeIndex === -1) {
         post.likes.push(userObjectId);
       } else {
@@ -221,53 +201,6 @@ class BlogService {
       return post;
     } catch (error) {
       throw new Error('Failed to toggle like');
-    }
-  }
-
-  async getSimilarPosts(postId: string) {
-    try {
-      const post = await Post.findById(postId);
-      if (!post) {
-        throw new Error('Post not found');
-      }
-
-      const allPosts = await Post.find({});
-
-      const similarPosts = await contentSimilarityService.findSimilarPosts({ id: post._id.toString(), title: post.title, content: post.content }, allPosts.map(p => ({ id: p._id.toString(), title: p.title, content: p.content })));
-      
-      const data = [];
-      for (const post of similarPosts) {
-        const p = await Post.findById(post.post.id);
-        if (p) {
-          data.push(p);
-        }
-      }
-      return data;
-    } catch (error) {
-      throw new Error('Failed to get similar posts');
-    }
-  }
-
-  async getPostRecommendations(postId: string) {
-    try {
-      const post = await Post.findById(postId);
-      if (!post) {
-        throw new Error('Post not found');
-      }
-
-      const allProducts = await Product.find({});
-      const similarProducts = await contentSimilarityService.findSimilarPosts({ id: post._id.toString(), title: post.title, content: post.content }, allProducts.map(p => ({ id: p._id.toString(), title: p.name, content: p.description })));
-      
-      const data = [];
-      for (const post of similarProducts) {
-        const p = await Product.findById(post.post.id);
-        if (p) {
-          data.push(p);
-        }
-      }
-      return data;
-    } catch (error) {
-      throw new Error('Failed to get post recommendations');
     }
   }
 }
