+++ frontend/app/(main)/instruction-blogs/[id]/page.tsx
@@ -3,12 +3,12 @@
 import { format } from 'date-fns'
 import Link from 'next/link'
 import { useEffect, useState } from 'react'
-import { FiArrowRight, FiCopy, FiEye, FiHeart, FiMessageCircle, FiShare2 } from 'react-icons/fi'
+import { FiArrowRight, FiCopy, FiEye, FiHeart, FiMessageCircle, FiSend, FiShare2, FiThumbsDown, FiThumbsUp } from 'react-icons/fi'
 import { useParams } from 'next/navigation'
 import { useSession } from 'next-auth/react'
 import axios from 'axios'
-import { toast } from "sonner"
-import Comments from './components/Comments'
+
+import { toast } from "sonner";
 
 interface SessionUser {
   id: string;
@@ -54,23 +54,75 @@ interface Post {
   isPublished: boolean
 }
 
-interface Product {
-  _id: string
-  name: string
-  description: string
-}
-
 export default function BlogViewPage() {
   const params = useParams()
-  const id = Array.isArray(params.id) ? params.id[0] : params.id || ''
+  const id = Array.isArray(params.id) ? params.id[0] : params.id
   const { data: session } = useSession()
   
   const [loading, setLoading] = useState(true)
   const [error, setError] = useState<string | null>(null)
+  const [newComment, setNewComment] = useState('')
+  const [replyTo, setReplyTo] = useState<string | null>(null)
   const [post, setPost] = useState<Post | null>(null)
   const [comments, setComments] = useState<Comment[]>([])
-  const [suggestedPosts, setSuggestedPosts] = useState<Post[]>([])
-  const [suggestedProducts, setSuggestedProducts] = useState<Product[]>([])
+  const [suggestedPosts] = useState<Post[]>([
+    {
+      _id: '2',
+      title: 'Healthy Living Tips',
+      content: 'Discover the best practices for maintaining a healthy lifestyle...',
+      thumbnail: 'https://images.unsplash.com/photo-1490645935967-10de6ba17061?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1000&q=80',
+      createdAt: '2024-03-19T15:30:00Z',
+      updatedAt: '2024-03-19T15:30:00Z',
+      category: 'Health',
+      author: {
+        _id: '',
+        name: '',
+        email: ''
+      },
+      views: 0,
+      likes: [],
+      comments: [],
+      isPublished: false
+    },
+    {
+      _id: '3',
+      title: 'The Future of Web Development',
+      content: 'Explore the latest trends and technologies shaping the future of web development...',
+      thumbnail: 'https://images.unsplash.com/photo-1498050108023-c5249f4df085?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1000&q=80',
+      createdAt: '2024-03-18T09:15:00Z',
+      updatedAt: '2024-03-18T09:15:00Z',
+      category: 'Technology',
+      author: {
+        _id: '',
+        name: '',
+        email: ''
+      },
+      views: 0,
+      likes: [],
+      comments: [],
+      isPublished: false
+    },
+    {
+      _id: '4',
+      title: 'Business Strategy in 2024',
+      content: 'Learn about the key business strategies that will drive success in 2024...',
+      thumbnail: 'https://images.unsplash.com/photo-1454165804606-c3d57bc86b40?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1000&q=80',
+      createdAt: '2024-03-17T14:20:00Z',
+      updatedAt: '2024-03-17T14:20:00Z',
+      category: 'Business',
+      author: {
+        _id: '',
+        name: '',
+        email: ''
+      },
+      views: 0,
+      likes: [],
+      comments: [],
+      isPublished: false
+    },
+  ])
+  const [page, setPage] = useState(1)
+  const [commentSubmitting, setCommentSubmitting] = useState(false)
 
   // Fetch post data
   useEffect(() => {
@@ -91,22 +143,8 @@ export default function BlogViewPage() {
 
     if (id) {
       fetchPost()
-      fetchSimilarPosts()
-      fetchPostRecommendations()
     }
   }, [id])
-
-  const fetchSimilarPosts = async () => {
-    const response = await axios.get(`${process.env.NEXT_PUBLIC_BASE_URL}/blog/${id}/similar`)
-    const thumbnailUrls = await Promise.all(response.data.data.map(async (post: Post) => await getThumbnailUrl(post.thumbnail)))
-    console.log(thumbnailUrls)
-    setSuggestedPosts(response.data.data.map((post: Post, index: number) => ({ ...post, thumbnail: thumbnailUrls[index] })))
-  }
-
-  const fetchPostRecommendations = async () => {
-    const response = await axios.get(`${process.env.NEXT_PUBLIC_BASE_URL}/blog/${id}/recommendations`)
-    setSuggestedProducts(response.data.data)
-  }
 
   const getThumbnailUrl = async (thumbnailPath: string): Promise<string> => {
     if (thumbnailPath.startsWith('http')) {
@@ -127,7 +165,7 @@ export default function BlogViewPage() {
   useEffect(() => {
     const fetchComments = async () => {
       try {
-        const response = await axios.get(`${process.env.NEXT_PUBLIC_BASE_URL}/comments/post/${id}?page=1&limit=10`)
+        const response = await axios.get(`${process.env.NEXT_PUBLIC_BASE_URL}/comments/post/${id}?page=${page}&limit=10`)
         setComments(response.data.data.comments)
       } catch (err) {
         console.error('Failed to load comments:', err)
@@ -137,9 +175,39 @@ export default function BlogViewPage() {
     if (id) {
       fetchComments()
     }
-  }, [id])
+  }, [id, page])
+
+  // Fetch suggested posts
+  useEffect(() => {
+    // const fetchSuggestedPosts = async () => {
+    //   try {
+    //     // Get posts from the same category if available
+    //     const category = post?.category
+    //     const url = category 
+    //       ? `${process.env.NEXT_PUBLIC_BASE_URL}/blog?page=1&limit=3&category=${encodeURIComponent(category)}`
+    //       : `${process.env.NEXT_PUBLIC_BASE_URL}/blog?page=1&limit=3`
+        
+    //     const response = await axios.get(url)
+        
+    //     // Filter out the current post
+    //     const filtered = response.data.data.posts.filter((p: Post) => p._id !== id)
+    //     setSuggestedPosts(filtered.slice(0, 3))
+    //   } catch (err) {
+    //     console.error('Failed to load suggested posts:', err)
+    //   }
+    // }
+
+    if (post) {
+      // fetchSuggestedPosts()
+    }
+  }, [post, id])
 
   const handleLike = async () => {
+    // if (!session) {
+    //   toast.error('Please sign in to like this post')
+    //   return
+    // }
+
     try {
       const response = await axios.post(`${process.env.NEXT_PUBLIC_BASE_URL}/blog/${id}/like`)
       
@@ -161,6 +229,198 @@ export default function BlogViewPage() {
     await navigator.clipboard.writeText(url)
     toast.success('Link copied to clipboard')
   }
+
+  const handleCommentLike = async (commentId: string) => {
+    // if (!session) {
+    //   toast.error('Please sign in to like this comment')
+    //   return
+    // }
+
+    try {
+      const response = await axios.post(`${process.env.NEXT_PUBLIC_BASE_URL}/comments/${commentId}/like`)
+      
+      // Update comments state with the updated comment
+      setComments(prev => {
+        return prev.map(comment => {
+          if (comment._id === commentId) {
+            return { ...comment, likes: response.data.data.likes }
+          }
+          
+          // Check if the comment is in replies
+          if (comment.replies && comment.replies.length > 0) {
+            const updatedReplies = comment.replies.map(reply => {
+              if (reply._id === commentId) {
+                return { ...reply, likes: response.data.data.likes }
+              }
+              return reply
+            })
+            
+            return { ...comment, replies: updatedReplies }
+          }
+          
+          return comment
+        })
+      })
+    } catch (err) {
+      toast.error('Failed to like comment')
+      console.error(err)
+    }
+  }
+
+  const handleCommentDislike = async (commentId: string) => {
+    // if (!session) {
+    //   toast.error('Please sign in to dislike this comment')
+    //   return
+    // }
+
+    try {
+      const response = await axios.post(`${process.env.NEXT_PUBLIC_BASE_URL}/comments/${commentId}/dislike`)
+      
+      // Update comments state with the updated comment
+      setComments(prev => {
+        return prev.map(comment => {
+          if (comment._id === commentId) {
+            return { ...comment, dislikes: response.data.data.dislikes }
+          }
+          
+          // Check if the comment is in replies
+          if (comment.replies && comment.replies.length > 0) {
+            const updatedReplies = comment.replies.map(reply => {
+              if (reply._id === commentId) {
+                return { ...reply, dislikes: response.data.dislikes }
+              }
+              return reply
+            })
+            
+            return { ...comment, replies: updatedReplies }
+          }
+          
+          return comment
+        })
+      })
+    } catch (err) {
+      toast.error('Failed to dislike comment')
+      console.error(err)
+    }
+  }
+
+  const handleSubmitComment = async (e: React.FormEvent) => {
+    e.preventDefault()
+    
+    // if (!session) {
+    //   toast.error('Please sign in to comment')
+    //   return
+    // }
+    
+    if (!newComment.trim()) return
+
+    try {
+      setCommentSubmitting(true)
+      
+      const commentData = {
+        content: newComment,
+        postId: id,
+        parentCommentId: replyTo || undefined
+      }
+      
+      const response = await axios.post(`${process.env.NEXT_PUBLIC_BASE_URL}/comments`, commentData)
+      const newCommentData = response.data.data
+      
+      // If it's a reply, add it to the parent comment's replies
+      if (replyTo) {
+        setComments(prev => {
+          return prev.map(comment => {
+            if (comment._id === replyTo) {
+              return {
+                ...comment,
+                replies: [...(comment.replies || []), newCommentData]
+              }
+            }
+            return comment
+          })
+        })
+      } else {
+        // If it's a top-level comment, add it to the comments list
+        setComments(prev => [newCommentData, ...prev])
+      }
+      
+      // Update post comment count in state
+      setPost(prev => {
+        if (!prev) return prev
+        return {
+          ...prev,
+          comments: [...prev.comments, newCommentData._id]
+        }
+      })
+      
+      toast.success(replyTo ? 'Reply posted successfully' : 'Comment posted successfully')
+      setNewComment('')
+      setReplyTo(null)
+    } catch (err) {
+      toast.error('Failed to post comment')
+      console.error(err)
+    } finally {
+      setCommentSubmitting(false)
+    }
+  }
+
+  const renderComment = (comment: Comment) => (
+    <div key={comment._id} className="space-y-4">
+      <div className="flex gap-4">
+        <img
+          src={`https://ui-avatars.com/api/?name=${encodeURIComponent(comment.author.name || '')}`}
+          alt={comment.author.name}
+          className="w-10 h-10 rounded-full"
+        />
+        <div className="flex-1">
+          <div className="bg-gray-50 rounded-lg p-4">
+            <div className="flex items-center gap-2 mb-2">
+              <span className="font-medium text-gray-900">{comment.author.name}</span>
+              <span className="text-sm text-gray-500">
+                {format(new Date(comment.createdAt), 'MMM d, yyyy')}
+              </span>
+            </div>
+            <p className="text-gray-700">{comment.content}</p>
+            <div className="flex items-center gap-4 mt-3">
+              <button
+                onClick={() => handleCommentLike(comment._id)}
+                className={`flex items-center gap-1 transition-colors ${
+                  comment.likes.includes((session?.user as SessionUser)?.id)
+                    ? 'text-blue-600'
+                    : 'text-gray-600 hover:text-blue-600'
+                }`}
+              >
+                <FiThumbsUp className="w-4 h-4" />
+                <span className="text-sm">{comment.likes.length}</span>
+              </button>
+              <button
+                onClick={() => handleCommentDislike(comment._id)}
+                className={`flex items-center gap-1 transition-colors ${
+                  session && session.user && comment.dislikes.includes((session.user as SessionUser).id)
+                    ? 'text-red-600'
+                    : 'text-gray-600 hover:text-red-600'
+                }`}
+              >
+                <FiThumbsDown className="w-4 h-4" />
+                <span className="text-sm">{comment.dislikes.length}</span>
+              </button>
+              <button
+                onClick={() => setReplyTo(comment._id)}
+                className="text-sm text-gray-600 hover:text-purple-600 transition-colors"
+              >
+                Reply
+              </button>
+            </div>
+          </div>
+          {comment.replies && comment.replies.length > 0 && (
+            <div className="ml-8 mt-4 space-y-4">
+              {comment.replies.map(reply => renderComment(reply))}
+            </div>
+          )}
+        </div>
+      </div>
+    </div>
+  )
 
   if (loading) {
     return <div className="flex justify-center items-center min-h-screen">
@@ -195,7 +455,7 @@ export default function BlogViewPage() {
               alt={'Avatar'}
               className="w-6 h-6 rounded-full"
             />
-            <span>{'Doctor'}</span>
+            <span>{'Avatar'}</span>
           </div>
           <span>•</span>
           <span>{format(new Date(post.createdAt), 'MMMM d, yyyy')}</span>
@@ -248,7 +508,80 @@ export default function BlogViewPage() {
       </div>
 
       {/* Comments Section */}
-      <Comments postId={id} initialComments={comments} />
+      <div className="space-y-8">
+        <h2 className="text-2xl font-bold text-gray-900">Comments</h2>
+        
+        {/* Comment Form */}
+        <form onSubmit={handleSubmitComment} className="flex gap-4">
+          <img
+            src={session && session.user ? `https://ui-avatars.com/api/?name=${encodeURIComponent((session.user as SessionUser).name || 'User')}` : 'https://ui-avatars.com/api/?name=Guest'}
+            alt={session && session.user ? (session.user as SessionUser).name || 'Your avatar' : 'Guest'}
+            className="w-10 h-10 rounded-full"
+          />
+          <div className="flex-1">
+            <textarea
+              value={newComment}
+              onChange={(e) => setNewComment(e.target.value)}
+              placeholder={"Write a reply..."}
+              className="w-full p-4 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-purple-500 focus:border-transparent resize-none"
+              rows={3}
+              disabled={commentSubmitting}
+            />
+            <div className="flex justify-end mt-2">
+              {replyTo && (
+                <button
+                  type="button"
+                  onClick={() => setReplyTo(null)}
+                  className="text-sm text-gray-600 hover:text-gray-900 mr-2"
+                  disabled={commentSubmitting}
+                >
+                  Cancel
+                </button>
+              )}
+              <button
+                type="submit"
+                className={`flex items-center gap-2 px-4 py-2 bg-purple-600 text-white rounded-lg transition-colors ${
+                  commentSubmitting ? 'opacity-50 cursor-not-allowed' : 'hover:bg-purple-700'
+                }`}
+                disabled={commentSubmitting}
+              >
+                {commentSubmitting ? (
+                  <>
+                    <span className="animate-spin h-4 w-4 border-2 border-white border-t-transparent rounded-full mr-2"></span>
+                    Submitting...
+                  </>
+                ) : (
+                  <>
+                    <FiSend className="w-4 h-4" />
+                    {replyTo ? 'Reply' : 'Comment'}
+                  </>
+                )}
+              </button>
+            </div>
+          </div>
+        </form>
+
+        {/* Comments List */}
+        <div className="space-y-8">
+          {comments.length > 0 ? (
+            comments.map(comment => renderComment(comment))
+          ) : (
+            <p className="text-gray-500 text-center py-4">No comments yet. Be the first to comment!</p>
+          )}
+        </div>
+        
+        {/* Load More Comments Button */}
+        {comments.length > 0 && (
+          <div className="flex justify-center">
+            <button
+              onClick={() => setPage(prev => prev + 1)}
+              className="px-4 py-2 border border-gray-300 rounded-lg text-gray-700 hover:bg-gray-50 transition-colors"
+            >
+              Load More Comments
+            </button>
+          </div>
+        )}
+      </div>
 
       {/* Suggested Posts */}
       <div className="mt-16 pt-8 border-t border-gray-200">
@@ -310,53 +643,6 @@ export default function BlogViewPage() {
             <p className="col-span-3 text-center text-gray-500 py-8">No suggested posts found</p>
           )}
         </div>
-
-      {/* Recommended Products Section */}
-      <div className="mt-12">
-        <div className="flex items-center justify-between mb-6">
-          <h2 className="text-2xl font-bold text-gray-900">Recommended Products</h2>
-          <Link href="/products" className="flex items-center text-purple-600 hover:text-purple-800 transition-colors">
-            View all products <FiArrowRight className="ml-2" />
-          </Link>
-        </div>
-        
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-          {suggestedProducts.length > 0 ? (
-            suggestedProducts.map((product) => (
-              <Link 
-                key={product._id} 
-                href={`/products/${product._id}`}
-                className="group"
-              >
-                <article className="bg-white rounded-lg shadow-sm border border-gray-200 overflow-hidden hover:shadow-md transition-shadow">
-                  <div className="relative h-40">
-                    <img
-                      src={`https://via.placeholder.com/300x200?text=${encodeURIComponent(product.name)}`}
-                      alt={product.name}
-                      className="w-full h-full object-cover group-hover:scale-105 transition-transform duration-300"
-                    />
-                  </div>
-                  <div className="p-4">
-                    <h3 className="font-semibold text-gray-900 mb-2 group-hover:text-purple-600 transition-colors">
-                      {product.name}
-                    </h3>
-                    <p className="text-gray-600 text-sm line-clamp-3">
-                      {product.description}
-                    </p>
-                    <div className="mt-4 flex justify-end">
-                      <span className="inline-flex items-center text-sm font-medium text-purple-600 group-hover:text-purple-800">
-                        Learn more <FiArrowRight className="ml-1 w-4 h-4" />
-                      </span>
-                    </div>
-                  </div>
-                </article>
-              </Link>
-            ))
-          ) : (
-            <p className="col-span-3 text-center text-gray-500 py-8">No recommended products found</p>
-          )}
-        </div>
-      </div>
       </div>
     </div>
   )
