+++ frontend/app/(private)/dashboard/blog/create-blog/page.tsx
@@ -115,9 +115,6 @@ export default function CreateBlog() {
             console.error('Error in form submission:', error);
         } finally {
             setIsSubmitting(false);
-            setTimeout(() => {
-                location.reload();
-            }, 1000)
         }
     };
