+++ frontend/package.json
@@ -38,7 +38,6 @@
     "@tiptap/pm": "^2.11.5",
     "@tiptap/react": "^2.11.5",
     "@tiptap/starter-kit": "^2.11.5",
-    "@types/lodash": "^4.17.16",
     "axios": "^1.7.9",
     "chart.js": "^4.4.8",
     "class-variance-authority": "^0.7.1",
@@ -46,7 +45,6 @@
     "date-fns": "^4.1.0",
     "framer-motion": "^12.4.7",
     "jspdf": "^3.0.0",
-    "lodash": "^4.17.21",
     "lucide-react": "^0.475.0",
     "next": "^15.3.1",
     "next-auth": "^4.24.11",
